Changes and added docker compose

+++ src/main/java/com/sprc_web_app/services/CityService.java
@@ -9,9 +9,12 @@ import com.sprc_web_app.model.entity.CountryEntity;
 import com.sprc_web_app.repositories.CityRepository;
 import com.sprc_web_app.repositories.CountryRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -23,12 +26,22 @@ public class CityService {
     private final CityMapper cityMapper;
 
     public CityIdResponse createCity(CityRequestDTO cityRequestDTO) {
+        boolean existsById = countryRepository.existsById(cityRequestDTO.getIdTara());
+        if (!existsById) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Country not found");
+        }
+
+        Optional<CityEntity> cityAux = cityRepository.findByIdTaraAndNume(cityRequestDTO.getIdTara(), cityRequestDTO.getNume());
+
+        if (!cityAux.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "Existing city name and country id");
+        }
 
 
         CityEntity cityEntity = cityMapper.mapCityRequestToEntity(cityRequestDTO);
         CountryEntity countryEntity = countryRepository.findById(cityRequestDTO.getIdTara()).orElseThrow();
         cityEntity.setCountry(countryEntity);
-        cityEntity = cityRepository.save(cityEntity);
+        cityEntity = cityRepository.saveAndFlush(cityEntity);
 
         return cityMapper.mapCityEntityToIdResponse(cityEntity);
     }
@@ -47,12 +60,16 @@ public class CityService {
         CityEntity cityEntity = cityRepository.findById(id).orElseThrow();
         cityEntity.setNume(cityRequestDTO.getNume());
         cityEntity.setCountry(countryRepository.findById(cityRequestDTO.getIdTara()).orElseThrow());
-        cityEntity = cityRepository.save(cityEntity);
+        cityEntity = cityRepository.saveAndFlush(cityEntity);
 
         return cityMapper.mapCityEntityToIdResponse(cityEntity);
     }
 
     public void deleteCity(Long id) {
+        if (!cityRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "City not found");
+        }
+
         cityRepository.deleteById(id);
     }
 }
